+++ donnees.h
@@ -1,8 +1,8 @@
 #ifndef donnees_h
 #define donnees_h
 
-float recup_L(float L);
-float recup_H(float H);
+double recup_L(float L);
+double recup_H(float H);
 int recup_M(float M);
 int recup_n(float n);
